worker/requirements.txt
# ─── Dev ──────────────────────────────────────────────────────────────────────
watchdog[watchmedo]      # Auto-reload worker on file change

# ─── HTTP Sidecar ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn==0.34.0

# ─── Core ─────────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1            # HTTP client (async)
python-dotenv==1.0.1

# ─── Scheduler ────────────────────────────────────────────────────────────────
apscheduler==3.10.4

# ─── Data Processing ──────────────────────────────────────────────────────────
pandas>=2.0.0
pandas-ta         # Technical indicators (RSI, MACD, Bollinger, etc.) - let pip resolve
numpy
vnstock           # Vietnam stock data — VCI source (no auth required)

# ─── AI / Embeddings ──────────────────────────────────────────────────────────
openai==1.68.2           # text-embedding-3-small

# ─── News Crawling ────────────────────────────────────────────────────────────
beautifulsoup4==4.13.3
feedparser==6.0.11       # RSS feeds
